+++ packages/fluentui/projects-test/src/createReactApp.ts
@@ -1,13 +1,13 @@
 import config from '@fluentui/scripts/config';
-import sh from '@fluentui/scripts/gulp/sh';
 import fs from 'fs-extra';
 import path from 'path';
 
 import {
   addResolutionPathsForProjectPackages,
   packProjectPackages,
-  createTempDir,
+  prepareTempDirs,
   log,
+  shEcho,
   performBrowserTest,
   prepareCreateReactApp,
 } from '@fluentui/scripts/projects-test';
@@ -23,11 +23,13 @@ export async function createReactApp() {
   const logger = log('test:projects:cra-ts');
   const scaffoldPath = config.paths.withRootAt(path.resolve(__dirname, '../assets/cra'));
 
-  const tmpDirectory = createTempDir('project-cra-');
-  logger(`✔️ Temporary directory was created: ${tmpDirectory}`);
+  const tempPaths = prepareTempDirs('project-cra-');
+  logger(`✔️ Temporary directories created under ${tempPaths.root}`);
+
   logger('STEP 1. Create test React project with TSX scripts..');
 
-  const testAppPath = config.paths.withRootAt(await prepareCreateReactApp(tmpDirectory, 'typescript', 'test-app'));
+  await prepareCreateReactApp(tempPaths, 'typescript');
+  const testAppPath = config.paths.withRootAt(tempPaths.testApp);
   logger(`Test React project is successfully created: ${testAppPath()}`);
 
   logger('STEP 2. Add Fluent UI dependency to test project..');
@@ -35,15 +37,16 @@ export async function createReactApp() {
   const packedPackages = await packProjectPackages(logger, config.paths.packages(), ['@fluentui/react-northstar']);
   await addResolutionPathsForProjectPackages(testAppPath());
 
-  await sh(`yarn add ${packedPackages['@fluentui/react-northstar']}`, testAppPath());
+  await shEcho(`yarn add ${packedPackages['@fluentui/react-northstar']}`, testAppPath());
   logger(`✔️ Fluent UI packages were added to dependencies`);
 
   logger("STEP 3. Reference Fluent UI components in test project's App.tsx");
   fs.copyFileSync(scaffoldPath('App.tsx'), testAppPath('src', 'App.tsx'));
 
   logger('STEP 4. Build test project..');
-  await sh(`yarn build`, testAppPath());
+  await shEcho(`yarn build`, testAppPath());
 
+  logger('STEP 5. Load the test app in the browser');
   await performBrowserTest(testAppPath('build'));
   logger(`✔️ Browser test was passed`);
 }