+++ packages/jest-serializer-merge-styles/jest.config.js
@@ -2,7 +2,7 @@ const path = require('path');
 const { createConfig } = require('@fluentui/scripts/jest/jest-resources');
 
 const config = createConfig({
-  snapshotSerializers: [path.resolve(__dirname)],
+  snapshotSerializers: [path.join(__dirname, 'src', 'index.ts')],
 });
 
 module.exports = config;