+++ packages/react-input/.storybook/preview.js
@@ -4,4 +4,11 @@ import * as rootPreview from '../../../.storybook/preview';
 export const decorators = [...rootPreview.decorators];
 
 /** @type {typeof rootPreview.parameters} */
-export const parameters = { ...rootPreview.parameters };
+export const parameters = {
+  ...rootPreview.parameters,
+  viewMode: 'story',
+  controls: {
+    disable: false,
+    expanded: false,
+  },
+};